Update to use the new target distance logic
This should be way more robust for large targets and max melee range.
The calc is now:
```
target.distance:sqrt() > self.melee_distance + self_mob.model_size + target.model_size - 0.1
```
The 0.1 is a correction for distance for smaller model enemies.

This calc is based on the calc in distanceplus, so it should now match
those calculations