Update melee distance and puller logic (#276)
* Updates the melee mode logic to always be accurate for max melee

This uses the calculation that the game uses which is:

2 + self.model_size + target.model_size where 2 is the melee
distance. I subtract a 0.2 from it for some edge cases.

* Doc comment fix

* Update puller.lua for better multi-pull functionality

This now selects the 1-max(n, 6)th target at random from the
list of mobs. A later patch will need to make that configurable.

This greatly reduces the chances of two Trusts targeting the same
target after a kill. This drastically improves speed in
situations like Odyssey and Dynamis where splitting up to fight
targets is optimal