Upgrade shooter to be a lot faster.
This updates shooter to be significantly faster by eliminating
the forced delay. This removes the forced delay. The actual
delay for shooting is a two-stage delay similar to JAs. The
first forced delay is the shooting animation. That can be
immediately followed by another ranged attack. However, following
with another ability like a WS requires a separate 1.0s delay.

This patch doesn't implement that, but significantly speeds up
shooting in all cases. Weaponskills do luckily retry, so the
overall change is minimal. In an ideal case we can add on more